Liten endring backend for raskere henting av grupper #deploy-backend-dev
