+++ apps/dolly-backend/src/main/java/no/nav/dolly/domain/resultset/entity/testgruppe/RsTestgruppePage.java
@@ -4,9 +4,12 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import no.nav.dolly.domain.jpa.Testgruppe;
 
 import java.util.ArrayList;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
 
 import static java.util.Objects.isNull;
 
@@ -21,11 +24,19 @@ public class RsTestgruppePage {
     private Integer pageSize;
     private Long antallElementer;
     private List<RsTestgruppe> contents;
+    private Set<Testgruppe> favoritter;
 
     public List<RsTestgruppe> getContents() {
         if (isNull(contents)) {
             contents = new ArrayList<>();
         }
         return contents;
+    }
+
+    public Set<Testgruppe> getFavoritter() {
+        if (isNull(favoritter)) {
+            favoritter = new HashSet<>();
+        }
+        return favoritter;
     }
 }
