+++ apps/dolly-backend/src/main/java/no/nav/dolly/provider/api/TestgruppeController.java
@@ -21,7 +21,6 @@ import no.nav.dolly.domain.resultset.entity.testgruppe.RsLockTestgruppe;
 import no.nav.dolly.domain.resultset.entity.testgruppe.RsOpprettEndreTestgruppe;
 import no.nav.dolly.domain.resultset.entity.testgruppe.RsTestgruppe;
 import no.nav.dolly.domain.resultset.entity.testgruppe.RsTestgruppeMedBestillingId;
-import no.nav.dolly.domain.resultset.entity.testgruppe.RsTestgruppeMedBestillingPage;
 import no.nav.dolly.domain.resultset.entity.testgruppe.RsTestgruppePage;
 import no.nav.dolly.service.BestillingService;
 import no.nav.dolly.service.SplittGruppeService;
@@ -126,7 +125,7 @@ public class TestgruppeController {
     @Cacheable(CACHE_GRUPPE)
     @GetMapping
     @Operation(description = "Hent testgrupper")
-    public RsTestgruppeMedBestillingPage getTestgrupper(
+    public RsTestgruppePage getTestgrupper(
             @RequestParam(value = "pageNo", required = false, defaultValue = "0") Integer pageNo, @RequestParam(value = "pageSize", required = false, defaultValue = "10") Integer pageSize, @RequestParam(name = "brukerId", required = false) String brukerId) {
         return testgruppeService.getTestgruppeByBrukerId(pageNo, pageSize, brukerId);
     }
